[refactor] Component Refactoring
