--- client/src/components/Meal/Main/style.ts
@@ -1,10 +1,0 @@
-import styled from '@emotion/styled';
-
-const MainContentsContainer = styled.div`
-  height: 100%;
-  display: flex;
-  width: 100%;
-  padding: 10px;
-`;
-
-export default MainContentsContainer;
