+++ client/src/pages/MainPage.tsx
@@ -1,21 +1,37 @@
-import React, { useState } from 'react';
+import React from 'react';
 import Header from '../components/Set/Header';
 import styled from '@emotion/styled';
-import Content from '../components/Set/Main';
-// import Test from '../components/Meal/Main/Content';
+import Input from '@set/Input';
+import Output from '@set/Output';
+import SaveButtons from '@set/SaveButtons';
 
 const MainContainer = styled.section`
   display: flex;
   flex-direction: column;
   height: 100%;
 `;
+const Main = styled.div`
+  height: 100%;
+  display: flex;
+`;
+const ContentsContainer = styled.div`
+  height: 100%;
+  display: flex;
+  width: 100%;
+  padding: 10px;
+`;
 
 function MainPage() {
   return (
     <MainContainer>
       <Header />
-      {/* <Test /> */}
-      <Content />
+      <Main>
+        <ContentsContainer>
+          <Input />
+          <Output />
+        </ContentsContainer>
+        <SaveButtons />
+      </Main>
     </MainContainer>
   );
 }
